+++ contracts/cyberpunk/tests/integration.rs
@@ -21,6 +21,7 @@ use cosmwasm_std::{from_binary, Empty, Env, Response};
 use cosmwasm_vm::testing::{
     execute, instantiate, mock_env, mock_info, mock_instance, mock_instance_with_gas_limit, query,
 };
+use std::time::SystemTime;
 
 use cyberpunk::msg::{ExecuteMsg, QueryMsg};
 
@@ -51,6 +52,35 @@ fn execute_argon2() {
     let expected = 8635688250000; // +/- 20%
     assert!(gas_used > expected * 80 / 100, "Gas used: {}", gas_used);
     assert!(gas_used < expected * 120 / 100, "Gas used: {}", gas_used);
+}
+
+// Test with
+// cargo integration-test debug_works -- --nocapture
+#[test]
+fn debug_works() {
+    let mut deps = mock_instance_with_gas_limit(WASM, 100_000_000_000_000);
+
+    let _res: Response =
+        instantiate(&mut deps, mock_env(), mock_info("admin", &[]), Empty {}).unwrap();
+
+    let msg = ExecuteMsg::Debug {};
+    let _res: Response = execute(&mut deps, mock_env(), mock_info("caller", &[]), msg).unwrap();
+
+    let start = SystemTime::now();
+    deps.set_debug_handler(move |msg, info| {
+        let gas = info.gas_remaining;
+        let runtime = SystemTime::now().duration_since(start).unwrap().as_micros();
+        eprintln!("{msg} (gas: {gas}, runtime: {runtime}µs)");
+    });
+
+    let msg = ExecuteMsg::Debug {};
+    let _res: Response = execute(&mut deps, mock_env(), mock_info("caller", &[]), msg).unwrap();
+
+    eprintln!("Unsetting debug handler. From here nothing is printed anymore.");
+    deps.unset_debug_handler();
+
+    let msg = ExecuteMsg::Debug {};
+    let _res: Response = execute(&mut deps, mock_env(), mock_info("caller", &[]), msg).unwrap();
 }
 
 #[test]