+++ packages/vm/src/environment.rs
@@ -1,8 +1,11 @@
 //! Internal details to be used by instance.rs only
 use std::borrow::{Borrow, BorrowMut};
+use std::marker::PhantomData;
 use std::ptr::NonNull;
+use std::rc::Rc;
 use std::sync::{Arc, RwLock};
 
+use derivative::Derivative;
 use wasmer::{HostEnvInitError, Instance as WasmerInstance, Memory, Val, WasmerEnv};
 use wasmer_middlewares::metering::{get_remaining_points, set_remaining_points, MeteringPoints};
 
@@ -79,11 +82,34 @@ impl GasState {
     }
 }
 
+/// Additional environmental information in a debug call.
+///
+/// The currently unused lifetime parameter 'a allows accessing referenced data in the debug implementation
+/// without cloning it.
+#[derive(Derivative)]
+#[derivative(Debug)]
+#[non_exhaustive]
+pub struct DebugInfo<'a> {
+    pub gas_remaining: u64,
+    // This field is just to allow us to add the unused lifetime parameter. It can be removed
+    // at any time.
+    #[doc(hidden)]
+    #[derivative(Debug = "ignore")]
+    pub(crate) __lifetime: PhantomData<&'a ()>,
+}
+
+// Unfortunately we cannot create an alias for the trait (https://github.com/rust-lang/rust/issues/41517).
+// So we need to copy it in a few places.
+//
+//                            /- BEGIN TRAIT                   END TRAIT \
+//                            |                                          |
+//                            v                                          v
+pub type DebugHandlerFn = dyn for<'a> Fn(/* msg */ &'a str, DebugInfo<'a>);
+
 /// A environment that provides access to the ContextData.
 /// The environment is clonable but clones access the same underlying data.
 pub struct Environment<A: BackendApi, S: Storage, Q: Querier> {
     pub api: A,
-    pub print_debug: bool,
     pub gas_config: GasConfig,
     data: Arc<RwLock<ContextData<S, Q>>>,
 }
@@ -96,7 +122,6 @@ impl<A: BackendApi, S: Storage, Q: Querier> Clone for Environment<A, S, Q> {
     fn clone(&self) -> Self {
         Environment {
             api: self.api,
-            print_debug: self.print_debug,
             gas_config: self.gas_config.clone(),
             data: self.data.clone(),
         }
@@ -110,13 +135,25 @@ impl<A: BackendApi, S: Storage, Q: Querier> WasmerEnv for Environment<A, S, Q> {
 }
 
 impl<A: BackendApi, S: Storage, Q: Querier> Environment<A, S, Q> {
-    pub fn new(api: A, gas_limit: u64, print_debug: bool) -> Self {
+    pub fn new(api: A, gas_limit: u64) -> Self {
         Environment {
             api,
-            print_debug,
             gas_config: GasConfig::default(),
             data: Arc::new(RwLock::new(ContextData::new(gas_limit))),
         }
+    }
+
+    pub fn set_debug_handler(&self, debug_handler: Option<Rc<DebugHandlerFn>>) {
+        self.with_context_data_mut(|context_data| {
+            context_data.debug_handler = debug_handler;
+        })
+    }
+
+    pub fn debug_handler(&self) -> Option<Rc<DebugHandlerFn>> {
+        self.with_context_data(|context_data| {
+            // This clone here requires us to wrap the function in Rc instead of Box
+            context_data.debug_handler.clone()
+        })
     }
 
     fn with_context_data_mut<C, R>(&self, callback: C) -> R
@@ -353,6 +390,7 @@ pub struct ContextData<S: Storage, Q: Querier> {
     storage_readonly: bool,
     call_depth: usize,
     querier: Option<Q>,
+    debug_handler: Option<Rc<DebugHandlerFn>>,
     /// A non-owning link to the wasmer instance
     wasmer_instance: Option<NonNull<WasmerInstance>>,
 }
@@ -365,6 +403,7 @@ impl<S: Storage, Q: Querier> ContextData<S, Q> {
             storage_readonly: true,
             call_depth: 0,
             querier: None,
+            debug_handler: None,
             wasmer_instance: None,
         }
     }
@@ -429,7 +468,7 @@ mod tests {
         Environment<MockApi, MockStorage, MockQuerier>,
         Box<WasmerInstance>,
     ) {
-        let env = Environment::new(MockApi::default(), gas_limit, false);
+        let env = Environment::new(MockApi::default(), gas_limit);
 
         let module = compile(CONTRACT, TESTING_MEMORY_LIMIT, &[]).unwrap();
         let store = module.store();